Call getFixedRoomMemberContent() when loading room members of a room
It fixes bad names for Empty room
